Introducing valueIsChanged in exp proxy interfaces/implementations
This new method is supposed to be called when a getted value is changed
(inline edit) or by set methods (internally) and when implemented
completely, will take care of values which must be changed upon an
explicit change, implicitly. For example writing to %ax explicitly,
will also change %eax implicitly.