INS disassembly was wrong in analysis routines. It's fixed
It seems (not documented or its documentation can not be found) that the
INS is dependent on some context which is present in instrumentation
methods and can not be moved to analysis routines. So calls to
INS_Disassemble method from analysis routines, return wrong results.
To resolve it, I instantiated string objects from disassembled
instruction within the instrumentation routines and trasnferred just
those string objects to analysis routines.