Always creating reg symbol for the largest enclosing register
While creating a symbol for a register which its size is less
than 64 bits, e.g. ax (a 16 bits register), a symbol for the
corresponding 64 bits register, i.e. rax, is created and value of
target register is calculated based on it, i.e. (sym(rax) & 0xFFFF).
This reduces the number of possible registers symbols.