Changing readRegisterContent to only read full-size regs using PIN API
The PIN_GetContextRegval function seems to segfault sometimes when it
is called to read a register which is not full-size (e.g. RAX is
full-size but EAX, AX, AH, and AL are not full-size).
Now the readRegisterContent calls PIN API using REG_FullRegName and then
parses its result to find the actually requested register's value.