Suppressing trivial constraints (the check is not correct in all cases)
All memory addresses which are not inside stack (including heap) are
assumed to be non-input. This assumption is not correct and should be
replaced with a more general code which detects output of syscalls and
main function's argc/argv and mark them as real symbols and left other
symbols as constant non-input values.
Till then, this simplification is useful for removing a lot of fake
memory symbols and removing related constraints.

Constraints who just contain non symbolic expressions are trivial (they
are always satisfied or changed to yet another constraint; but may not
change their true/false state without being changed themselves).
While constraints are being added to the constraints tree, the trivial
constraints are ignored.