Updating entire tree of mem-expressions in valueIsChanged
Previously, the subtree rooted at the changed expression would be
updated in addition to the path from the changed expressions towards the
root. With this commit, the subtree rooted at the neighbor is
checked/updated too. This renews whole of the tree. However, it is not
required when the neighbor expression is already created (as it is
already in synch with all other expressions in its subtree). But there
is no straight method to detect this state yet.