Writing a test case to check constraints having BITVECTOR_UREM_TOTAL
This commit updates interfaces to make this thing usable/testable too.

Some notes:
  ∘ Both remainder and division operations are not defined on zero
divisor.
  ∘ For example both 3 / 0 and 3 % 0 raise "Floating point exception."
  ∘ Consequently, these cases will cause program crash.
  ∘ When we start checking for program crashes, this case can be modeled
by adding a constraint dictating that divisor is zero.
  ∘ So, in current implementation and also after supporting app crash
scenarios, whenever the division or remainder operations are seen in an
expression and/or constraint, the divisor was non-zero for sure (if it
was zero, it would cause crash and the expression wouldn't include
it; the constraint which modeles divisor as zero does not contain
the division or remainder by zero in itself too).
  ∘ Thus it is safe (and it is required) to add a (divisor!=zero)
constraint to all constraints/expressions who contain a
division/remainder operation.
  ∘ Also ITE constructs that check divisor against zero in CVC4
simplified expressionmust be replaced with the divisor!=zero case.
  ∘ This test shows a simplified expresion to check whether it handles
above two items:
    1. adding divisor!=zero to the constraint
    2. removing the path of ITE that is for zero divisor.