Updating Flags class to maintain state of flags separately
Now arg expressions and the operation which was performed between them
are kept in Flags class and whenever a check is performed, state of
flags (which may be in synch with those args and operation or may be
    set/clear separately) will be inspected to instantiate the
appropriate constraint. Non-conditional cases are detected and null is
returned as a sign that no constraint is required to be satisfied in
order to follow the executed path.