Fixing add-with-carry operation
ADC adds its two operands and the CF (which indicates the carry of the
    previous operation) and stores result in its destination argument
and if this operation itself produces carry, the CF will be set again.
It is common to call ADD and then ADC to capture the carry of ADD
instruction and use it in the ADC operation.

NOTE: This is different from one's complement sum where the carry of
the current operation is summed again. One's complement sum of eax and
ebx can be calculated using two instructions
like this: add eax, ebx; adc eax, 0x0