Bugfix in Operator instantiation and called overridden virtual method
When a virtual method which is overridden in this child class is invoked
within the constructor of the parent class, the parent's implementation
will be called instead of the overridden child's implementation. Because
within the constructor, the child class is not constructed completely
and its implementation pointer is not registered yet.