Fixing a concrete value overflow during `(Z / c1) >> c2` simplification
During the simplification, (Z/c1)*c1 was calculated to reverse the
effect of division in order to apply the right-shift to the Z directly.
However, if the precision of concrete value of (Z/c1) was lower than Z
(e.g. due to moving to smaller register), then Z==Z/c1*c1 won't be hold
necessarily.