Creating memory symbols using shift instead of division
Memory is currently managed in 64 bits units and sub-symbols are
calculated by division and bitwise-and of the enclosing 64 bits main
symbols. However, value of those sub-symbols may be simplified in Twinner
and re-inputted to Twintool as bitwise-shift (instead of division).
In order to obtain the same constraints, syntactically, while driving
the Twintool through a specified path, the original sub-symbol should be
instantiated using bitwise-shifts too.