Limiting validity checks in TreeNode::addConstraint to some depth
Without validity checks, it is possible that tautological constraints
get added to the ETG. This won't cause any logical problem, but it
reduces the performance when SMTS should check a lot of unsatisfiable
execution paths and eliminate them one by one. On the other hand, it
increases the performance of adding the constraints at the first time.