Limit transferred mem (as temp symbol from past seg) to its read size
Previously the 128-bits aligned address containing the transferred mem
was calculated and whole of the previously-loaded content were copied
as temp symbol to the current seg. This fails whenever the
requested mem content is up-to-date but has an out-of-date neighbor.
Not transferring neighbors, allows those neighbors to be replaced
with a new real symbol and other up-to-date neighbors to be copied
as temp symbols in future requests (if any).