Marking state change as expected for old segments
After failed lazy load in an old segment, an unexpected change is
reported which causes the trace try-to-get method to stop.
However, if the failed lazy load was performed on an old segment, it
should be reverted to an expected change scenario allowing
the new segment to create its own symbol.