Supporting two separate offsets for argc and argv
Useful for when argument are passed by registers and so cannot
be located in correct order on the stack. Note that it is not possible
to get them from registers directly since the analysis routine
is not necessarily called at the beginning of the main function
itself (e.g. in the -endpoints mode).