Improving BitwiseAndOperator::deepSimplify code
After propagating the simplification over sub expressions, if result
had at least three tokens, the simplification could be continued
and when there were fewer tokens, CAN_NOT_SIMPLIFY was returned
to simply add the `& operand` formula to the remaining single
token (or unary operator).

However, if the remaining token is a constant, it can be simplified
more. The new code returns RESTART_SIMPLIFICATION to check for
constant being of the remaining token.